refactor: Rename and remove unuse file
